--- .editorconfig
@@ -1,16 +1,0 @@
-# Editor configuration, see https://editorconfig.org
-root = true
-
-[*]
-charset = utf-8
-indent_style = space
-indent_size = 2
-insert_final_newline = true
-trim_trailing_whitespace = true
-
-[*.ts]
-quote_type = double
-
-[*.md]
-max_line_length = off
-trim_trailing_whitespace = false
