--- lefthook.yml
@@ -1,7 +1,0 @@
-pre-commit:
-  commands:
-    check:
-      glob: "*.{js,ts,cjs,mjs,d.cts,d.mts,jsx,tsx,json,jsonc}"
-      root: "react/"
-      run: bunx biome check --write --error-on-warnings {staged_files}
-      stage_fixed: true
