+++ package.json
@@ -10,6 +10,8 @@
     "check": "biome check --error-on-warnings --write ."
   },
   "dependencies": {
+    "@base-ui/react": "^1.3.0",
+    "@fontsource-variable/raleway": "^5.2.8",
     "@radix-ui/react-accordion": "^1.2.12",
     "@radix-ui/react-dialog": "^1.1.15",
     "@radix-ui/react-dropdown-menu": "^2.1.16",
@@ -21,6 +23,7 @@
     "react": "^19.2.4",
     "react-dom": "^19.2.4",
     "react-router": "^7.13.1",
+    "shadcn": "^4.1.2",
     "supertokens-auth-react": "^0.51.2",
     "tailwind-merge": "^3.5.0"
   },
