--- react/example.env
@@ -1,2 +1,0 @@
-VITE_API_DOMAIN=""
-VITE_WEBSITE_DOMAIN=""
