--- react/tsconfig.node.json
@@ -1,11 +1,0 @@
-{
-  "compilerOptions": {
-    "composite": true,
-    "skipLibCheck": true,
-    "module": "ESNext",
-    "moduleResolution": "bundler",
-    "allowSyntheticDefaultImports": true,
-    "strictNullChecks": true
-  },
-  "include": ["vite.config.ts"]
-}
